open source comp additions
